Moved the CORS request file to a new folder, Shared
