+++ LibraryAccess/src/shared/createCORSRequest.js
